requirements.txt
pandas
openai
psycopg
python-dotenv
timescale-vector
instructor
anthropic

transformers
torch